# api/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/films":{"get":{"tags":["Film Controller"],"summary":"Get all films","operationId":"getAll_1","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmResponseDto"}}}}}}},"post":{"tags":["Film Controller"],"summary":"Create film","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmNewDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmResponseDto"}}}}}}},"/directors":{"get":{"tags":["Director Controller"],"summary":"Get all directors","operationId":"getAll_2","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DirectorResponseDto"}}}}}}},"post":{"tags":["Director Controller"],"summary":"Create director","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectorNewDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DirectorResponseDto"}}}}}}},"/tracktime":{"get":{"tags":["TrackTimeLog Controller"],"summary":"Get all track time logs","operationId":"getAll","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrackTimeLogDto"}}}}}}}},"/tracktime/{methodName}":{"get":{"tags":["TrackTimeLog Controller"],"summary":"Get all track time logs by method name","operationId":"getAllByMethodName","parameters":[{"name":"methodName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrackTimeLogDto"}}}}}}}},"/tracktime/stat/{methodName}":{"get":{"tags":["TrackTimeLog Controller"],"summary":"Get full stats by method name","operationId":"getStatByMethodName","parameters":[{"name":"methodName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrackTimeLogStatDto"}}}}}}},"/tracktime/average/{methodName}":{"get":{"tags":["TrackTimeLog Controller"],"summary":"Get average duration by method name","operationId":"getAverageDurationByMethodName","parameters":[{"name":"methodName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrackTimeLogAverageDurationDto"}}}}}}},"/films/{filmId}":{"get":{"tags":["Film Controller"],"summary":"Get film by id","operationId":"getByFilmId","parameters":[{"name":"filmId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FilmResponseDto"}}}}}},"delete":{"tags":["Film Controller"],"summary":"Delete film","operationId":"delete","parameters":[{"name":"filmId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"No Content"}}}},"/films/director/{directorId}":{"get":{"tags":["Film Controller"],"summary":"Get all films by director id","operationId":"getAllByDirectorId","parameters":[{"name":"directorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmResponseDto"}}}}}}}},"/directors/{directorId}":{"get":{"tags":["Director Controller"],"summary":"Get director by id","operationId":"getById","parameters":[{"name":"directorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DirectorResponseDto"}}}}}},"delete":{"tags":["Director Controller"],"summary":"Delete director","operationId":"delete_1","parameters":[{"name":"directorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"No Content"}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"status":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 EARLY_HINTS","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"reason":{"type":"string"},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}}}},"FilmNewDto":{"required":["description","directorId","name"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"rating":{"type":"number","format":"double"},"directorId":{"type":"integer","format":"int64"}},"description":"Dto for creating new film"},"DirectorResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"birthYear":{"type":"integer","format":"int32"}},"description":"Dto for response to director request"},"FilmResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"rating":{"type":"number","format":"double"},"director":{"$ref":"#/components/schemas/DirectorResponseDto"}},"description":"Dto for response to film request"},"DirectorNewDto":{"required":["name"],"type":"object","properties":{"name":{"type":"string"},"birthYear":{"type":"integer","format":"int32"}},"description":"Dto for creating new director"},"TrackTimeLogDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"methodName":{"type":"string"},"methodArgs":{"type":"string"},"isAsync":{"type":"boolean"},"timestamp":{"type":"string","format":"date-time"},"duration":{"type":"integer","format":"int64"}},"description":"Dto for response to track time log request"},"TrackTimeLogStatDto":{"type":"object","properties":{"methodName":{"type":"string"},"callCount":{"type":"integer","format":"int64"},"minimumDuration":{"type":"integer","format":"int64"},"maximumDuration":{"type":"integer","format":"int64"},"averageDuration":{"type":"number","format":"double"}},"description":"Dto for response to full stats by method name request"},"TrackTimeLogAverageDurationDto":{"type":"object","properties":{"methodName":{"type":"string"},"averageDuration":{"type":"number","format":"double"}},"description":"Dto for response to average duration by method name request"}}}}